version 0.1.0
- fix bug with countries
- added a screen for collection without words to learn
- added a deletion confirmation prompt for learning cards
- screenshots and demo for the readme file
- context field for words is multiline now

+++ src/pages/overview/Overview.tsx
@@ -4,9 +4,10 @@ import {deleteCollection, getCollections} from '../../shared/model/storage';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import {AddCollectionButton} from './ui/AddCollectionButton';
 import {useQuery} from '../../shared/lib/useQuery';
-import {showDeleteConfirmationAlert} from './ui/showDeleteConfirmationAlert';
 import {PagesStackProps} from '../../shared/model/navigator';
 import {CollectionItems} from './ui/CollectionItems';
+import {showConfirmationAlert} from '../../shared/lib/message';
+import {translate} from '../../shared/lib/i18n';
 
 export const Overview = ({
   navigation,
@@ -29,7 +30,10 @@ export const Overview = ({
             })
           }
           onDelete={() =>
-            showDeleteConfirmationAlert()
+            showConfirmationAlert(
+              translate('collection_deletion'),
+              translate('collection_deletion_message'),
+            )
               .then(() => deleteCollection(collection))
               .finally(() => refetch())
           }
